expose camel/{id}/registry endpoint

+++ src/main/java/nl/ordina/bertkoor/camelcase/repo/CamelRegistry.java
@@ -1,10 +1,9 @@
 package nl.ordina.bertkoor.camelcase.repo;
 
+import nl.ordina.bertkoor.camelcase.model.RegistryData;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface CamelRegistry extends CrudRepository<RegistryData, BigInteger> {
+public interface CamelRegistry extends CrudRepository<RegistryData, Long> {
 }
